refcator and added class
